refactor(issue-288): normalizar Descripcion nula y cerrar cobertura de los tres niveles
- Resuelve el bloqueo heredado: el fixture DetalleEsperado de
  SolicitarProgramacionTurnoCommandHandlerTests conservaba el placeholder ""
  que el propio test-writer documento como pendiente de la fase verde. Ahora
  lleva el texto real que produce el ToString() del tipo rico (CA-2).

- Corrige un defecto contra la "consecuencia asumida" del issue: los eventos
  anteriores a #288 no traen el campo y STJ dejaba null en una propiedad
  declarada string no anulable (verificado empiricamente), no "" como declara
  el issue. Los tres DTOs planos lo normalizan a cadena vacia.

- Agrega el guardrail de retro-compatibilidad que lo cubre y extiende el
  round-trip de portabilidad por el bus a los tres niveles (turno, franja y
  sub-franja), que antes solo cubria dos.

- Refuerza CatalogoTurnosTests con el caso compuesto (turno partido con
  descanso), donde el ToString() del turno concatena ordinarias y arrastra
  los labels .resx.

- Limpia cruft: comentarios "queda en rojo hasta que el implementer..." ya
  obsoletos en los tres archivos de coherencia y casts (int)0 redundantes.

+++ tests/Bitakora.ControlAsistencia.PrivateEvents.Tests/Programacion/DetalleFranjaOrdinariaIgualdadTests.cs
@@ -13,10 +13,10 @@ namespace Bitakora.ControlAsistencia.PrivateEvents.Tests.Programacion;
 public class DetalleFranjaOrdinariaIgualdadTests : IgualdadTestBase<DetalleFranjaOrdinaria>
 {
     private static DetalleSubFranja Descanso() =>
-        new(new TimeOnly(12, 0), new TimeOnly(13, 0), (int)0, (int)0, "");
+        new(new TimeOnly(12, 0), new TimeOnly(13, 0), 0, 0, "");
 
     private static DetalleSubFranja Extra() =>
-        new(new TimeOnly(17, 0), new TimeOnly(19, 0), (int)0, (int)0, "");
+        new(new TimeOnly(17, 0), new TimeOnly(19, 0), 0, 0, "");
 
     protected override DetalleFranjaOrdinaria CrearInstancia() =>
         new(new TimeOnly(8, 0), new TimeOnly(17, 0), 0, [Descanso()], [Extra()], "");
